refactor(compose): the company view's three prose lanes get their own file
The rebase onto main pushed serveroptions.go past the 500-line cap. The three
grounded-prose lane options are one concept: they share the decision of what a
reader sees when no model is available, and the dossier's and growth fit's both
rebuild the same handler set, so either may run first.
