+++ backend/internal/compose/serveroptions.go
@@ -15,8 +15,6 @@ import (
 
 	"github.com/jackc/pgx/v5/pgxpool"
 
-	"github.com/gradionhq/margince/backend/internal/compose/orgbrief"
-	"github.com/gradionhq/margince/backend/internal/compose/orgdossier"
 	"github.com/gradionhq/margince/backend/internal/modules/activities"
 	"github.com/gradionhq/margince/backend/internal/modules/approvals"
 	"github.com/gradionhq/margince/backend/internal/modules/customfields"
@@ -450,66 +448,5 @@ func WithExtractor(extractor extraction.Extractor) Option {
 func WithBrief(brain completer) Option {
 	return func(s *Server, _ *pgxpool.Pool) {
 		s.WithL2Ranker(brain, s.log)
-	}
-}
-
-// WithAccountBrief binds the summarize lane both of the company view's
-// grounded-prose surfaces are written by — the standing brief and the
-// prepared "Ask Margince" questions — and the routing version that
-// identifies the binding in every cached brief's fingerprint.
-//
-// Without it both serve their deterministic floor rather than failing: a
-// role that runs no model still answers the endpoints, and generated_by
-// tells the reader which of the two they have. routingVersion rides the
-// fingerprint so re-pointing this lane rewrites cached briefs instead of
-// leaving text attributed to a model that no longer writes it.
-func WithAccountBrief(brain completer, routingVersion string) Option {
-	return func(s *Server, pool *pgxpool.Pool) {
-		s.orgBriefSvc = orgbrief.NewService(pool, s.org360Svc, s.peopleStore, brain, routingVersion, time.Now)
-		s.orgBriefHandlers = orgbrief.NewHandlers(s.orgBriefSvc, s.sorDispatch.isOverlay)
-	}
-}
-
-// WithCompanyDossier binds the lane that writes what a company IS, and the
-// routing version that identifies the binding in every cached dossier's
-// fingerprint.
-//
-// Unlike the growth fit's, this lane is an improvement rather than a
-// precondition: the floor already describes the company from the same fields,
-// one restated value per sentence. What the model adds is prose a person reads
-// before a call instead of a list they skim.
-//
-// It rebuilds the shared handler set from BOTH services for the reason
-// WithGrowthFit does — replacing that set while remembering only one service
-// would leave the other endpoint answering from a service the Server no longer
-// holds. Either option may run first.
-func WithCompanyDossier(brain completer, routingVersion string) Option {
-	return func(s *Server, pool *pgxpool.Pool) {
-		s.orgDossierSvc = orgdossier.NewService(pool, s.peopleStore, brain, routingVersion, time.Now)
-		s.orgDossierHandlers = orgdossier.NewHandlers(
-			s.orgDossierSvc, s.orgGrowthFitSvc, s.sorDispatch.isOverlay)
-	}
-}
-
-// WithGrowthFit binds the lane that judges how well a company fits what we
-// sell, and the routing version that identifies the binding in every cached
-// assessment's fingerprint.
-//
-// This is the one company-view lane whose absence CHANGES THE ANSWER rather
-// than only its prose. The dossier's floor still describes a company; growth
-// fit's floor abstains by DOSS-PARAM-7, because grading is not a restatement of
-// recorded values. So an unwired deployment serves "here is what I would need
-// to know", labelled deterministic, and never a band nobody stands behind.
-//
-// The dossier service is rebound alongside it rather than rebuilt, because the
-// two share one handler set: replacing that set while remembering only one
-// service would leave the other endpoint answering from a service the Server no
-// longer holds.
-func WithGrowthFit(brain completer, routingVersion string) Option {
-	return func(s *Server, pool *pgxpool.Pool) {
-		s.orgGrowthFitSvc = orgdossier.NewGrowthFitService(
-			pool, s.peopleStore, offeringConfirmed(s.peopleStore), brain, routingVersion, time.Now)
-		s.orgDossierHandlers = orgdossier.NewHandlers(
-			s.orgDossierSvc, s.orgGrowthFitSvc, s.sorDispatch.isOverlay)
 	}
 }
